# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JobMatch API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /saved:
    get:
      operationId: listSaved
      tags: [matches]
      summary: List saved/bookmarked jobs
      responses:
        "200":
          description: Saved job IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedJob"
    post:
      operationId: saveJob
      tags: [matches]
      summary: Save a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedJobInput"
      responses:
        "201":
          description: Job saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedJob"
  /saved/{jobId}:
    delete:
      operationId: unsaveJob
      tags: [matches]
      summary: Remove a saved job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Job removed from saved
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SessionToken
  schemas:
...
    Job:
      type: object
      required: [id, title, company, description, location, skills, jobType, source, scrapedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        logoUrl:
          type: ["string", "null"]
          description: URL du logo de l'entreprise
        description:
          type: string
        location:
          type: string
        remote:
          type: boolean
        skills:
          type: array
          items:
            type: string
        jobType:
          type: string
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        source:
          type: string
        sourceUrl:
          type: ["string", "null"]
        postedAt:
          type: ["string", "null"]
        experienceYears:
          type: ["integer", "null"]
        sector:
          type: ["string", "null"]
        diplomaRequired:
          type: ["string", "null"]
        scrapedAt:
          type: string
...
    SavedJob:
      type: object
      required: [id, jobId, savedAt]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        savedAt:
          type: string
        job:
          $ref: "#/components/schemas/Job"
    SavedJobInput:
      type: object
      required: [jobId]
      properties:
        jobId:
          type: integer
...